Cleanup & Can now have screen sizes of any shape or size!
You can now have window sizes in *Ghasp* even numbers! How magical I know!

Also some code has been cleaned up and stuff. Fun I know

+++ src/GameInfo/Environment/Blocks/BlockBase.java
@@ -20,7 +20,7 @@ public abstract class BlockBase {
         blockType = BlockTypeEnum.TEST_FLOOR;
     }
 
-    public abstract void renderBlock(Canvas canvas, GraphicsContext gc, int x, int y, int renderLayer);
+    public abstract void renderBlock(Canvas canvas, GraphicsContext gc, double x, double y, int renderLayer);
 
 
     /**
